[class] : #10 자기 아이디만 탈퇴 가능하도록

+++ src/main/java/com/prj1/controller/MemberController.java
@@ -4,6 +4,7 @@ import com.prj1.domain.Member;
 import com.prj1.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.eclipse.tags.shaded.org.apache.xpath.operations.Mod;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -50,8 +51,10 @@ public class MemberController {
 
     // 회원 탈퇴 기능
     @PostMapping("remove")
-    public String remove(Integer id) {
-        service.remove(id);
+    public String remove(Integer id, Authentication auth) {
+        if (service.hasAccess(id, auth)) {
+            service.remove(id);
+        }
         return "redirect:/member/signup";
     }
